OEL-1574: Move selectSlimOption to dedicated trait.
